renamed SdkWs source
